refactor: reorganize goal-related hooks into mutations and queries for better structure and clarity
