+++ src/components/GoalsComponents/GoalSublist/GoalSublist.tsx
@@ -12,11 +12,11 @@ import DeletedGoals from "@pages/GoalsPage/components/DeletedGoals";
 import ArchivedGoals from "@pages/GoalsPage/components/ArchivedGoals";
 import GoalItemSummary from "@src/common/GoalItemSummary/GoalItemSummary";
 import AvailableGoalHints from "@pages/GoalsPage/components/AvailableGoalHints";
-import { useGetGoalById } from "@src/hooks/api/Goals/useGetGoalById";
+import { useGetGoalById } from "@src/hooks/api/Goals/queries/useGetGoalById";
 import { useGetSharedWMGoalsArchived } from "@src/hooks/api/SharedWMGoals/useGetSharedWMGoalsArchived";
-import { useGetArchivedGoals } from "@src/hooks/api/Goals/useGetArchivedGoals";
+import { useGetArchivedGoals } from "@src/hooks/api/Goals/queries/useGetArchivedGoals";
 import { useGetContactByPartnerId } from "@src/hooks/api/Contacts/useGetContactByPartnerId";
-import { useGetDeletedGoals } from "@src/hooks/api/Goals/useGetDeletedGoals";
+import { useGetDeletedGoals } from "@src/hooks/api/Goals/queries/useGetDeletedGoals";
 import { useParams } from "react-router-dom";
 import GoalsList from "../GoalsList";
 import GoalHistory from "./components/GoalHistory";
